+++ private/GRAPHQL_MULTIPART_REQUEST_SPEC_URL.js
@@ -1,12 +1,12 @@
 'use strict';
 
 /**
- * Official [GraphQL multipart request spec](https://github.com/jaydenseric/graphql-multipart-request-spec)
+ * [GraphQL multipart request spec](https://github.com/jaydenseric/graphql-multipart-request-spec)
  * URL. Useful for error messages, etc.
  * @kind constant
- * @name SPEC_URL
+ * @name GRAPHQL_MULTIPART_REQUEST_SPEC_URL
  * @type {string}
  * @ignore
  */
-exports.SPEC_URL =
+module.exports =
   'https://github.com/jaydenseric/graphql-multipart-request-spec';